Batch apply simple suggestions from code review

+++ stubs/networkx/networkx/algorithms/approximation/traveling_salesman.pyi
@@ -11,7 +11,8 @@ def simulated_annealing_tsp(
     init_cycle,
     weight: str = "weight",
     source: Incomplete | None = None,
-    temp: int = 100,
+    # docstring says int, but it can be a float and does become a float mid-equation if alpha is also a float
+    temp: float = 100,
     move: str = "1-1",
     max_iterations: int = 10,
     N_inner: int = 100,
@@ -23,7 +24,7 @@ def threshold_accepting_tsp(
     init_cycle,
     weight: str = "weight",
     source: Incomplete | None = None,
-    threshold: int = 1,
+    threshold: float = 1,
     move: str = "1-1",
     max_iterations: int = 10,
     N_inner: int = 100,
