+++ stubs/networkx/networkx/generators/geometric.pyi
@@ -1,15 +1,15 @@
 from _typeshed import Incomplete
 
-def geometric_edges(G, radius, p: int = 2): ...
+def geometric_edges(G, radius, p: float = 2): ...
 def random_geometric_graph(
-    n, radius, dim: int = 2, pos: Incomplete | None = None, p: int = 2, seed: Incomplete | None = None
+    n, radius, dim: int = 2, pos: Incomplete | None = None, p: float = 2, seed: Incomplete | None = None
 ): ...
 def soft_random_geometric_graph(
     n,
     radius,
     dim: int = 2,
     pos: Incomplete | None = None,
-    p: int = 2,
+    p: float = 2,
     p_dist: Incomplete | None = None,
     seed: Incomplete | None = None,
 ): ...
@@ -32,7 +32,8 @@ def waxman_graph(
     metric: Incomplete | None = None,
     seed: Incomplete | None = None,
 ): ...
-def navigable_small_world_graph(n, p: int = 1, q: int = 1, r: int = 2, dim: int = 2, seed: Incomplete | None = None): ...
+# docstring marks p as int, but it still works with floats. So I think it's better for consistency
+def navigable_small_world_graph(n, p: float = 1, q: int = 1, r: float = 2, dim: int = 2, seed: Incomplete | None = None): ...
 def thresholded_random_geometric_graph(
     n,
     radius,
@@ -40,6 +41,6 @@ def thresholded_random_geometric_graph(
     dim: int = 2,
     pos: Incomplete | None = None,
     weight: Incomplete | None = None,
-    p: int = 2,
+    p: float = 2,
     seed: Incomplete | None = None,
 ): ...
